+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Create player silently (do not start music fater player created)
 * Add shuffle playlist option after playlist repeated
+* Improve command parameters checking
+* Add force_play and force_pause to pause_playback command to force player state
+* Add volume parameter to pause_playback to change player volume at once
+* Update event returns internal state instead of Gst state
 
 ### Fixed